backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# Copy dependencies from original thinkr app
numpy
langchain
langchain-google-genai
langchain-google-vertexai
langchain-google-community
google-cloud-aiplatform
google-cloud-bigquery
pandas-gbq
unstructured[pdf]
